Extract and use HelpCallout component
