+++ packages/mobile/src/screens/edit-track-screen/screens/RemixSettingsScreen.tsx
@@ -26,6 +26,7 @@ import { useColor } from 'app/utils/theme'
 
 import { FormScreen, RemixTrackPill } from '../components'
 import type { RemixOfField } from '../types'
+import { HelpCallout } from 'app/components/help-callout/HelpCallout'
 
 const { getTrack, getUser, getStatus } = remixSettingsSelectors
 const { fetchTrack, fetchTrackSucceeded, reset } = remixSettingsActions
@@ -75,24 +76,11 @@ const useStyles = makeStyles(({ palette, spacing, typography }) => ({
     fontSize: typography.fontSize.large
   },
   changeAvailability: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: spacing(16),
-    paddingVertical: spacing(2),
-    paddingHorizontal: spacing(4),
-    backgroundColor: palette.neutralLight9,
-    borderWidth: 1,
-    borderColor: palette.neutralLight7,
-    borderRadius: spacing(2)
+    marginBottom: spacing(16)
   },
   changeAvailabilityText: {
     flexDirection: 'row',
     flexWrap: 'wrap'
-  },
-  questionIcon: {
-    marginRight: spacing(4),
-    width: spacing(5),
-    height: spacing(5)
   }
 }))
 
@@ -109,7 +97,6 @@ const descriptionProps: TextProps = {
 export const RemixSettingsScreen = () => {
   const isPremiumContentEnabled = useIsPremiumContentEnabled()
   const styles = useStyles()
-  const neutral = useColor('neutral')
   const [{ value: remixOf }, , { setValue: setRemixOf }] =
     useField<RemixOfField>('remix_of')
   const [{ value: remixesVisible }, , { setValue: setRemixesVisible }] =
@@ -233,20 +220,21 @@ export const RemixSettingsScreen = () => {
     >
       <View>
         <View style={styles.setting}>
-          {isPremium && (
-            <View style={styles.changeAvailability}>
-              <IconQuestionCircle style={styles.questionIcon} fill={neutral} />
-              <View style={styles.changeAvailabilityText}>
-                <Text>{messages.changeAvailbilityPrefix}</Text>
-                <Text>
-                  {isCollectibleGated
-                    ? messages.collectibleGated
-                    : messages.specialAccess}
-                </Text>
-                <Text>{messages.changeAvailbilitySuffix}</Text>
-              </View>
-            </View>
-          )}
+          {isPremium ? (
+            <HelpCallout
+              style={styles.changeAvailability}
+              content={
+                <View style={styles.changeAvailabilityText}>
+                  <Text>{messages.changeAvailbilityPrefix}</Text>
+                  <Text>
+                    {isCollectibleGated
+                      ? messages.collectibleGated
+                      : messages.specialAccess}
+                  </Text>
+                  <Text>{messages.changeAvailbilitySuffix}</Text>
+                </View>
+              } />
+          ) : null}
           <View style={styles.option}>
             <Text {...labelProps}>
               {isPremiumContentEnabled
